Create rightshift_by_rotating_integer.c
C program to develop a function rightshift(x,n) in C that returns the value of the integer x rotated to the right by n bit positions as an unsigned integer. Invoke the function from the main with different values for x and n and print the results with suitable headings.